Use separate module for the client config
